move the singletons to a single file
