+++ packages/shared.test/src/stubs/ontologies/legacyFullOntology.ts
@@ -14,14 +14,32 @@
  * limitations under the License.
  */
 
+import invariant from "tiny-invariant";
+import { FauxDataStore } from "../../FauxFoundry/FauxDataStore.js";
+import { FauxFoundry } from "../../FauxFoundry/FauxFoundry.js";
 import { FauxOntology } from "../../FauxFoundry/FauxOntology.js";
 import { registerLazyActions } from "../actions.js";
 import { BarInterface, FooInterface } from "../interfaces.js";
+import {
+  employee1 as employee_John_50030,
+  employee2 as employee_Jane_50031,
+  employee3,
+  employee4withDerived,
+  employee50050,
+  employeeFailsStrict,
+  nycOffice,
+  objectWithAllPropertyTypes1,
+  objectWithAllPropertyTypesEmptyEntries,
+  travisPlayer,
+} from "../objects.js";
 import { objectTypesWithLinkTypes } from "../objectTypesWithLinkTypes.js";
 import { queryTypes } from "../queryTypes.js";
 import { fooSpt } from "../spts.js";
 import { defaultOntologyMetadata } from "./defaultOntologyMetadata.js";
 
+//
+// Setup the ontology
+//
 export const legacyFullOntology: FauxOntology = new FauxOntology(
   defaultOntologyMetadata,
 );
@@ -40,3 +58,46 @@ legacyFullOntology.registerInterfaceType(BarInterface);
 legacyFullOntology.registerInterfaceType(FooInterface);
 
 legacyFullOntology.registerSharedPropertyType(fooSpt);
+
+//
+// Setup the data store
+//
+export const legacyFauxDataStore: FauxDataStore = new FauxDataStore(
+  legacyFullOntology,
+);
+legacyFauxDataStore.registerObject(employee_John_50030);
+legacyFauxDataStore.registerObject(employee_Jane_50031);
+legacyFauxDataStore.registerObject(employee3);
+legacyFauxDataStore.registerObject(employee4withDerived);
+legacyFauxDataStore.registerObject(employeeFailsStrict);
+legacyFauxDataStore.registerObject(employee50050);
+legacyFauxDataStore.registerObject(nycOffice);
+legacyFauxDataStore.registerObject(objectWithAllPropertyTypes1);
+legacyFauxDataStore.registerObject(objectWithAllPropertyTypesEmptyEntries);
+legacyFauxDataStore.registerObject(travisPlayer);
+legacyFauxDataStore.registerLink(
+  employee_John_50030,
+  "lead",
+  employee_Jane_50031,
+  "peeps",
+);
+legacyFauxDataStore.registerLink(
+  employee3,
+  "lead",
+  employee_Jane_50031,
+  "peeps",
+);
+invariant(
+  legacyFauxDataStore.getLinks(
+    "Employee",
+    employee_Jane_50031.__primaryKey,
+    "peeps",
+  ).length === 2,
+);
+
+//
+// Setup the faux foundry
+//
+export const fauxFoundry: FauxFoundry = new FauxFoundry();
+fauxFoundry.registerOntology(legacyFullOntology);
+fauxFoundry.setDataStore(legacyFullOntology.apiName, legacyFauxDataStore);
